chore: Restructure files
